+++ src/Farkle/CSharp/DesigntimeFarkleExtensions.fs
@@ -32,7 +32,18 @@ type ProductionBuilderExtensions =
 /// <summary>Extension methods for the <see cref="DesigntimeFarkle{TResult}"/> type.</summary>
 type DesigntimeFarkleExtensions =
     [<Extension>]
-    /// <summary>Builds a <see cref="DesigntimeFarkle{TResult}"/> into a <see cref="RuntimeFarkle{TResult}"/>.</summary>
+    /// <summary>Casts a <see cref="DesigntimeFarkle"/>
+    /// into a <see cref="DesigntimeFarkle{Object}"/></summary>
+    /// <remarks>Useful for setting metadata to untyped designtime Farkles.
+    /// The object it would return is undefined.</remarks>
+    static member Cast df = DesigntimeFarkle.cast df
+    [<Extension>]
+    /// <summary>Changes the name of a <see cref="DesigntimeFarkle{TResult}"/>.<summary>
+    /// <remarks>Useful for diagnostic purposes.</remarks>
+    static member Rename(df: DesigntimeFarkle<'TResult>, name) = DesigntimeFarkle.rename name df
+    [<Extension>]
+    /// <summary>Builds a <see cref="DesigntimeFarkle{TResult}"/>
+    /// into a <see cref="RuntimeFarkle{TResult}"/>.</summary>
     static member Build (df: DesigntimeFarkle<'TResult>) = RuntimeFarkle.build df
     [<Extension>]
     /// <summary>Builds a <see cref="DesigntimeFarkle"/> into a syntax-checking